feat: extract shared user constants to src/utils/userConstants.ts
